api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
faiss-cpu==1.7.4
numpy==1.24.4
httpx==0.25.0
python-dotenv==1.0.0
sentence-transformers==2.2.2 